Rename Page constants to Route
The routes are decoupled from the pages that they render. Some pages
may have multiple routes. Thats why RouteNames and RoutePaths is
a more appropriate name for a constant
